backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.129.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn[standard]==0.40.0
httpx==0.28.1
numpy==2.2.3
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.20
bcrypt==4.2.1
PyJWT==2.9.0
piper-tts>=1.4.0
edge-tts>=6.1.0
Pillow>=10.0.0
